third_party/blink/web_tests/external/wpt/css/css-display/parsing/tentative/display-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['display'] = "inline grid-lanes" should set the property value
  assert_equals: serialization should be canonical expected "inline grid-lanes" but got "inline-grid-lanes"
[FAIL] e.style['display'] = "grid-lanes inline" should set the property value
  assert_equals: serialization should be canonical expected "inline grid-lanes" but got "inline-grid-lanes"
Harness: the test ran to completion.
